+++ src/vscode/dropProvider/_types.ts
@@ -1,0 +1,14 @@
+import * as vscode from 'vscode';
+import { PowerBIDragMIMEType } from '../treeviews/PowerBIApiDragAndDropController';
+
+export interface PowerBIDaxDrop {
+	daxDrop: string;
+}
+
+export interface PowerBIApiDrop {
+	apiDrop: string;
+}
+
+export interface PowerBITmslDrop {
+	tmslDrop: string;
+}