+++ src/vscode/treeviews/workspaces/PowerBIDatasetTable.ts
@@ -4,15 +4,16 @@ import { PowerBIWorkspaceTreeItem } from './PowerBIWorkspaceTreeItem';
 import { Helper, UniqueId } from '../../../helpers/Helper';
 import { ThisExtension } from '../../../ThisExtension';
 import { PowerBIApiService } from '../../../powerbi/PowerBIApiService';
-import { PROCESSING_TYPES, PowerBIDataset } from './PowerBIDataset';
+import { PowerBIDataset } from './PowerBIDataset';
 import { iPowerBIDatasetColumnStatistics, iPowerBIDatasetDMV, iPowerBIDatasetRefreshableObject } from '../../../powerbi/DatasetsAPI/_types';
 import { PowerBIDatasetTables } from './PowerBIDatasetTables';
 import { PowerBIDatasetTableColumns } from './PowerBIDatasetTableColumns';
 import { PowerBIDatasetTableMeasures } from './PowerBIDatasetTableMeasures';
 import { PowerBIDatasetTablePartitions } from './PowerBIDatasetTablePartitions';
+import { PowerBIApiDrop, PowerBIDaxDrop, PowerBITmslDrop } from '../../dropProvider/_types';
 
 // https://vshaxe.github.io/vscode-extern/vscode/TreeItem.html
-export class PowerBIDatasetTable extends PowerBIWorkspaceTreeItem {
+export class PowerBIDatasetTable extends PowerBIWorkspaceTreeItem implements PowerBIDaxDrop, PowerBITmslDrop, PowerBIApiDrop{
 	private _columnStatistics: iPowerBIDatasetColumnStatistics[] = [];
 
 	constructor(
@@ -139,17 +140,35 @@ export class PowerBIDatasetTable extends PowerBIWorkspaceTreeItem {
 		return children;
 	}
 
+	// DAX Drop
+	get daxDrop(): string {
+		return `'${this.name}'`;
+	}
+
+	// TMSL Drop
+	get tmslDrop(): string {
+		return JSON.stringify(this.refreshableObject, null, 4);
+	}
+
+	// API Drop
+	get apiDrop(): string {
+		return JSON.stringify(this.refreshableObject, null, 4);
+	}
+
 	// DatasetTable-specific funtions
 	get refreshableObject(): iPowerBIDatasetRefreshableObject {
 		return { table: this.name };
 	}
 
 	public async refresh(): Promise<void> {
+		ThisExtension.TreeViewWorkspaces.doMultiselectAction("REFRESH");
+		/*
 		const isOnDedicatedCapacity = this.dataset.workspace.isPremiumCapacity;
 
 		await PowerBIDataset.refreshById(this.groupId.toString(), this.dataset.id, isOnDedicatedCapacity, [this.refreshableObject]);
 
 		await Helper.delay(1000);
 		ThisExtension.TreeViewWorkspaces.refresh(this.dataset, false);
+		*/
 	}
 }